+++ LaborVolumeCalculator/Models/Dictionary/StandardModulesUsingRate.cs
@@ -10,12 +10,12 @@ namespace LaborVolumeCalculator.Models.Dictionary
         {
         }
 
-        public StandardModulesUsingRate(string name, float value) : base(name)
+        public StandardModulesUsingRate(string name, double value) : base(name)
         {
             Value = value;
         }
 
         [DisplayName("Значение")]
-        public float Value { get; set; }
+        public double Value { get; set; }
     }
 }
